cross_domain: split mod.rs into separate files
Easier to reason about each object does.  Essentially,
there's a context (main thread), worker thread, and
shared state (which may spin it's own threads).
